Refactor to a single editor
