--- components/JSEditor/style.css
@@ -1,10 +1,0 @@
-button {
-  border: 3px solid var(--color-border);
-  border-radius: 4px;
-  min-height: 5vh;
-  width: 80px;
-}
-
-button:hover {
-  background-color: var(--color-active-highlight);
-}
