--- components/MarkdownEditor/MarkdownEditor.js
@@ -1,29 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import PropTypes from 'prop-types';
-import path from 'path';
-
-import css from './style.css';
-
-function MarkdownEditor({ file, write }) {
-  const [value, setValue] = useState('');
-
-  useEffect(() => {
-    (async () => {
-      setValue(await file.text());
-    })();
-  }, [file]);
-
-  return (
-    <div className={css.editor}>
-      <div className={css.title}>{path.basename(file.name)}</div>
-      <div className={css.content}>{value}</div>
-    </div>
-  );
-}
-
-MarkdownEditor.propTypes = {
-  file: PropTypes.object,
-  write: PropTypes.func
-};
-
-export default MarkdownEditor;
