+++ components/TextEditor/Toggle.js
